Renamed from RaindropPHP to Round Eights
Argh. The project started out as commonPHP, but I decided that was too... common. Marketing would be difficult with a name like that. So, I set out for a new name. It's an important and difficult process, as simple as it seems. Because I live in a rainy region and I liked the way a raindrop metaphorically lends itself to a coding library, I went with RaindropPHP. A few weeks ago, Mozilla announced their own project called Raindrop. While I realize the two projects could coexist with the same name, I am not interested in competing. So, another name change. Round Eights. In truth, I like this name much more than the first two. During the Victorian period, Round Eights was one of a few well known stage combat routines. Actors would recycle these routines to quickly and reliably stage fights. As before, this provides a strong metaphore and ties in to one of my hobbies. I hope you like it as much as I do.